+++ src/gas/traders/plugin/api.ts
@@ -53,4 +53,5 @@ export interface Signal {
 export enum SignalType {
   None,
   Buy,
+  Manual,
 }