+++ src/web/components/Settings.tsx
@@ -6,6 +6,7 @@ import {
   Avatar,
   Box,
   Button,
+  Chip,
   Divider,
   FormControl,
   FormControlLabel,
@@ -15,12 +16,11 @@ import {
   Link,
   MenuItem,
   Select,
+  Slider,
   Stack,
   Switch,
   TextField,
   Typography,
-  Slider,
-  Chip,
 } from "@mui/material";
 import { circularProgress, ScriptApp } from "./Common";
 import {
@@ -178,6 +178,7 @@ export function Settings(params: {
                 style={{ marginRight: 10 }}
               />
               Auto-stop (Market strength)
+              {cfg.TradingAutoStopped && ` (Now: Paused)`}
             </Typography>
             <Slider
               sx={{ ml: 1, width: `95%` }}
@@ -191,6 +192,16 @@ export function Settings(params: {
                 label: 10 * i,
               }))}
               valueLabelDisplay="auto"
+              onChangeCommitted={(e, [min, max]: number[]) => {
+                setCfg({
+                  ...cfg,
+                  TradingAutoStopped:
+                    // inverse the auto stop value when edge values selected
+                    min === 0 && max === 100
+                      ? !cfg.TradingAutoStopped
+                      : cfg.TradingAutoStopped,
+                });
+              }}
               onChange={(e, [min, max]: number[]) => {
                 setCfg({
                   ...cfg,
@@ -205,9 +216,9 @@ export function Settings(params: {
             <FormHelperText>
               This setting controls when to pause trading based on the market
               {` `}
-              <b>strength</b>. For example, if the range is 10-60, trading
+              <b>strength</b>. For example, if the range is 30-60, trading
               starts when the strength reaches 60 or more and stops when it
-              drops below 10. It resumes only after the strength reaches 60
+              drops below 30. It resumes only after the strength reaches 60
               again.
             </FormHelperText>
           </FormControl>